Making csv and json response reusable
